WIP: dumping threads on uncaught exceptions
Building ontop of last-resort pipeline exceptions here:
https://github.com/bazelbuild/bazel-buildfarm/pull/1309

Like jstack but better because we goveren when they are dumped and here
we do it prior to interruption.

A few categories of exceptions I've encountered are more obscure and
async in nature so this helps.  Not a silver bullet because one of the
issues is caused by events several cycles before the exception.